+++ src/core/map.cpp
@@ -22,8 +22,6 @@ namespace mas
                 grid_config.size       = {static_cast<float>(map_config_.grid_size),
                                           static_cast<float>(map_config_.grid_size)};
                 grid_config.fill_color = sf::Color::White;
-                // grid_config.outline_color     = sf::Color::Black;
-                // grid_config.outline_thickness = 1;
                 grids_[i][j].setGridConfig(grid_config);
             }
         }
@@ -51,7 +49,7 @@ namespace mas
         }
     }
 
-    sf::Vector2i Map::getGridIndex(const sf::Vector2i& position) const
+    sf::Vector2i Map::getGridIndex(const sf::Vector2f& position) const
     {
         return sf::Vector2i(position.x / map_config_.grid_size,
                             position.y / map_config_.grid_size);
